Initial code for to_bytes, from_bytes, exif_to_gps, and gps_to_exif
